+++ ApartmentService/src/main/java/com/dwellora/repository/ApartmentRepository.java
@@ -3,6 +3,7 @@ package com.dwellora.repository;
 import com.dwellora.entity.Apartment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApartmentRepository extends JpaRepository<Apartment, Integer> {
-
-}
+/**
+ * Repository interface for managing {@link Apartment} entities.
+ */
+public interface ApartmentRepository extends JpaRepository<Apartment, Long> {}
