+++ ApartmentService/src/main/java/com/dwellora/service/ApartmentService.java
@@ -4,16 +4,33 @@ import com.dwellora.dto.ApartmentRequestDTO;
 import com.dwellora.dto.ApartmentResponseDTO;
 import java.util.List;
 
+/**
+ * Service interface for managing apartment operations.
+ */
 public interface ApartmentService {
 
+    /**
+     * Adds a new apartment.
+     */
     ApartmentResponseDTO addApartment(ApartmentRequestDTO dto);
 
+    /**
+     * Retrieves all apartments.
+     */
     List<ApartmentResponseDTO> getAllApartments();
 
-    ApartmentResponseDTO getApartmentById(Integer apartmentId);
+    /**
+     * Retrieves an apartment by its ID.
+     */
+    ApartmentResponseDTO getApartmentById(Long apartmentId);
 
-    ApartmentResponseDTO updateApartment(
-            Integer apartmentId, ApartmentRequestDTO dto);
+    /**
+     * Updates an existing apartment by its ID.
+     */
+    ApartmentResponseDTO updateApartment(Long apartmentId, ApartmentRequestDTO dto);
 
-    void deleteApartment(Integer apartmentId);
+    /**
+     * Deletes an apartment by its ID.
+     */
+    void deleteApartment(Long apartmentId);
 }
