+++ OnboardingService/src/main/java/com/dwellora/service/OnboardingService.java
@@ -4,15 +4,18 @@ import com.dwellora.dto.OnboardingRequestDTO;
 import com.dwellora.dto.OnboardingResponseDTO;
 import java.util.List;
 
+/**
+ * Service interface for handling onboarding requests.
+ */
 public interface OnboardingService {
 
     OnboardingResponseDTO createRequest(OnboardingRequestDTO request);
 
     List<OnboardingResponseDTO> getPendingRequests();
 
-    OnboardingResponseDTO approveRequest(Integer requestId);
+    OnboardingResponseDTO approveRequest(Long requestId);
 
-    OnboardingResponseDTO rejectRequest(Integer requestId);
+    OnboardingResponseDTO rejectRequest(Long requestId);
 
     List<OnboardingResponseDTO> getAllRequests();
 }
